feat(components): organize/declutter shared components folder part 2
