+++ packages/blockchain-wallet-v4-frontend/src/components/Clipboard/CopyClipboard/template.js
