+++ packages/blockchain-wallet-v4-frontend/src/components/Clipboard/CopyClipboardButton/index.tsx
