+++ packages/blockchain-wallet-v4-frontend/src/components/Clipboard/CopyClipboardButton/template.tsx
