+++ packages/blockchain-wallet-v4-frontend/src/layouts/Public/components/Footer/index.tsx
