+++ packages/blockchain-wallet-v4-frontend/src/modals/Btc/RequestBtc/template.shareLink.js
@@ -7,7 +7,7 @@ import styled from 'styled-components'
 import { Button, Text, TextGroup } from 'blockchain-info-components'
 import { Exchange } from 'blockchain-wallet-v4/src'
 import CoinDisplay from 'components/Display/CoinDisplay'
-import CopyClipboard from 'components/CopyClipboard'
+import CopyClipboard from 'components/Clipboard/CopyClipboard'
 import QRCodeWrapper from 'components/QRCode/Wrapper'
 
 const Details = styled.div`
