+++ packages/blockchain-wallet-v4-frontend/src/modals/QRCode/template.js
@@ -14,7 +14,7 @@ import {
   TooltipHost,
   TooltipIcon
 } from 'blockchain-info-components'
-import CopyClipboard from 'components/CopyClipboard'
+import CopyClipboard from 'components/Clipboard/CopyClipboard'
 
 const QRCodeContainer = styled.div`
   display: flex;
