+++ packages/blockchain-wallet-v4-frontend/src/modals/Xlm/RequestXlm/template.success.js
@@ -16,7 +16,7 @@ import {
 } from 'components/Form'
 import { model } from 'data'
 import { required } from 'services/FormHelper'
-import CopyClipboard from 'components/CopyClipboard'
+import CopyClipboard from 'components/Clipboard/CopyClipboard'
 import QRCodeWrapper from 'components/QRCode/Wrapper'
 
 const AddressContainer = styled.div`
