+++ packages/blockchain-wallet-v4-frontend/src/scenes/loading.public.tsx
@@ -1,9 +1,9 @@
 import { FormattedMessage } from 'react-intl'
-import { SpinningLoader, Text } from 'blockchain-info-components'
 import React from 'react'
 import styled from 'styled-components'
 
-import Header from 'layouts/Public/Header'
+import { SpinningLoader, Text } from 'blockchain-info-components'
+import Header from 'layouts/Public/components/Header'
 import media from 'services/ResponsiveService'
 
 interface Props {}
